requirements.txt
beautifulsoup4==4.12.3
selenium==4.21.0
openai==1.30.5
tqdm
gradio>=4.0.0
pandas>=2.0.0
python-dateutil==2.8.2
matplotlib==3.8.2
numpy>=1.24.0
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.3
scikit-learn==1.6.1
textblob==0.17.1
tweepy==4.14.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0